Implements custom Cache for Interpreters
In order to reuse interpreters generated when training models with
Rasa NLU, a custom caching class has been implemented following the
architecture of functools.lru_cache but allowing the AsyncTrainer
class to store models in the cache.

Training status is persisted in the AsyncTrainer class and can be queried
through the status endpoint.

Model name is now expected in path